# L24-mcp-server-demo/openapi.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"tags":[{"name":"Admin APIs","description":"Endpoints for company, seller and category administration"},{"name":"Customer APIs","description":"Endpoints for product browsing and order lifecycle actions"},{"name":"Seller APIs","description":"Endpoints for seller operations like product management and order processing"}],"paths":{"/api/seller/products/{productId}":{"put":{"tags":["Seller APIs"],"summary":"Update product","description":"Updates an existing product by product id.","operationId":"updateProduct","parameters":[{"name":"productId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateProductRequestDto"}}},"required":true},"responses":{"400":{"description":"Validation error"},"200":{"description":"Product updated successfully"},"404":{"description":"Product, company, or category not found"}}}},"/api/seller/orders/{orderId}/accept":{"put":{"tags":["Seller APIs"],"summary":"Accept order","description":"Changes order status from PLACED to ACCEPTED.","operationId":"acceptOrder","parameters":[{"name":"orderId","in":"path","description":"Order id","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"404":{"description":"Order not found"},"200":{"description":"Order accepted successfully"},"400":{"description":"Order is not in PLACED status"}}}},"/api/customer/order/{id}/submit":{"put":{"tags":["Customer APIs"],"summary":"Submit order","description":"Changes order status from DRAFT to PLACED.","operationId":"submitOrder","parameters":[{"name":"id","in":"path","description":"Order id","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"400":{"description":"Order is not in DRAFT status"},"404":{"description":"Order not found"},"200":{"description":"Order submitted successfully"}}}},"/api/seller/product":{"post":{"tags":["Seller APIs"],"summary":"Create product","description":"Creates a new product for a company and category.","operationId":"createProduct","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateProductRequestDto"}}},"required":true},"responses":{"400":{"description":"Validation error"},"404":{"description":"Company or category not found"},"201":{"description":"Product created successfully"}}}},"/api/customer/order-item":{"post":{"tags":["Customer APIs"],"summary":"Add order item","description":"Adds item to user's active DRAFT order, or creates a new DRAFT order and adds it.","operationId":"createOrderItem","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateOrderItemRequestDto"}}},"required":true},"responses":{"201":{"description":"Order item added successfully"},"400":{"description":"Invalid request or stock issue"},"404":{"description":"User or product not found"}}}},"/api/admin/seller":{"post":{"tags":["Admin APIs"],"summary":"Create seller","description":"Creates a seller user for an existing company.","operationId":"createSeller","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateSellerRequestDto"}}},"required":true},"responses":{"404":{"description":"Company not found"},"400":{"description":"Validation error or duplicate email"},"201":{"description":"Seller created successfully"}}}},"/api/admin/company":{"post":{"tags":["Admin APIs"],"summary":"Create company","description":"Creates a company and its primary user in one request.","operationId":"createCompany","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateCompanyRequestDto"}}},"required":true},"responses":{"400":{"description":"Validation error or duplicate values"},"201":{"description":"Company created successfully"}}}},"/api/admin/category":{"post":{"tags":["Admin APIs"],"summary":"Create category","description":"Creates a new product category.","operationId":"createCategory","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateCategoryRequestDto"}}},"required":true},"responses":{"400":{"description":"Validation error or duplicate category name"},"201":{"description":"Category created successfully"}}}},"/api/seller/orders/pending":{"get":{"tags":["Seller APIs"],"summary":"Get pending orders","description":"Fetches all PLACED orders for a given company id.","operationId":"getPendingOrders","parameters":[{"name":"companyId","in":"query","description":"Company id","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"404":{"description":"Company not found"},"200":{"description":"Pending orders fetched successfully"}}}},"/api/customer/products":{"get":{"tags":["Customer APIs"],"summary":"Search products by keyword","description":"Returns products where keyword matches name or description.","operationId":"getProducts","parameters":[{"name":"keyword","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Products fetched successfully"},"400":{"description":"Invalid keyword"}}}},"/api/customer/order/{id}":{"get":{"tags":["Customer APIs"],"summary":"Get order details","description":"Fetches order summary, total, and all order-item details by order id.","operationId":"getOrderDetails","parameters":[{"name":"id","in":"path","description":"Order id","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"404":{"description":"Order not found"},"200":{"description":"Order fetched successfully"}}}},"/api/admin/sellers":{"get":{"tags":["Admin APIs"],"summary":"Get all sellers (paginated)","description":"Returns paginated list of users having SELLER role.","operationId":"getAllSellers","parameters":[{"name":"page","in":"query","description":"Zero-based page index","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","description":"Page size","required":false,"schema":{"type":"integer","format":"int32","default":10}},{"name":"sortBy","in":"query","description":"Sort field","required":false,"schema":{"type":"string","default":"id"}},{"name":"direction","in":"query","description":"Sort direction: asc or desc","required":false,"schema":{"type":"string","default":"asc"}}],"responses":{"200":{"description":"Sellers fetched successfully"}}}},"/api/admin/seller/{id}":{"delete":{"tags":["Admin APIs"],"summary":"Delete seller","description":"Deletes seller by id if seller exists and is not configured as a primary company user.","operationId":"deleteSeller","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"204":{"description":"Seller deleted successfully"},"400":{"description":"Invalid operation for this user"},"404":{"description":"Seller not found"}}}}},"components":{"schemas":{"UpdateProductRequestDto":{"required":["categoryId","companyId","name","price","stock"],"type":"object","properties":{"name":{"type":"string"},"description":{"type":"string"},"imageUrl":{"type":"string"},"price":{"minimum":0.0,"exclusiveMinimum":true,"type":"number"},"stock":{"type":"integer","format":"int32"},"companyId":{"type":"integer","format":"int64"},"categoryId":{"type":"integer","format":"int64"},"isActive":{"type":"boolean"}}},"CreateProductRequestDto":{"required":["categoryId","companyId","name","price","stock"],"type":"object","properties":{"name":{"type":"string"},"description":{"type":"string"},"imageUrl":{"type":"string"},"price":{"minimum":0.0,"exclusiveMinimum":true,"type":"number"},"stock":{"type":"integer","format":"int32"},"companyId":{"type":"integer","format":"int64"},"categoryId":{"type":"integer","format":"int64"},"isActive":{"type":"boolean"}}},"CreateOrderItemRequestDto":{"required":["productId","quantity","userId"],"type":"object","properties":{"userId":{"type":"integer","format":"int64"},"productId":{"type":"integer","format":"int64"},"quantity":{"minimum":1,"type":"integer","format":"int32"}}},"CreateSellerRequestDto":{"required":["companyId","email","name","password"],"type":"object","properties":{"companyId":{"type":"integer","format":"int64"},"name":{"type":"string"},"email":{"type":"string"},"password":{"type":"string"}}},"CreateCompanyRequestDto":{"required":["name","number","primaryUser"],"type":"object","properties":{"name":{"type":"string"},"number":{"type":"string"},"isActive":{"type":"boolean"},"primaryUser":{"$ref":"#/components/schemas/PrimaryUserDto"}}},"PrimaryUserDto":{"required":["email","name","password"],"type":"object","properties":{"name":{"type":"string"},"email":{"type":"string"},"password":{"type":"string"},"role":{"type":"string","enum":["ADMIN","SELLER","CUSTOMER"]}}},"CreateCategoryRequestDto":{"required":["name"],"type":"object","properties":{"name":{"type":"string"},"description":{"type":"string"}}}}}}