# L24-mcp-server-demo/src/main/java/com/example/L13_minor_project_01/dto/CreateProductRequestDto.java
package com.example.L13_minor_project_01.dto;

import jakarta.validation.constraints.DecimalMin;
import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.NotNull;
import jakarta.validation.constraints.PositiveOrZero;
import lombok.Getter;
import lombok.Setter;

import java.math.BigDecimal;

@Getter
@Setter
public class CreateProductRequestDto {

    @NotBlank
    private String name;

    private String description;

    private String imageUrl;

    @NotNull
    @DecimalMin(value = "0.0", inclusive = false)
    private BigDecimal price;

    @NotNull
    @PositiveOrZero
    private Integer stock;

    @NotNull
    private Long companyId;

    @NotNull
    private Long categoryId;

    private Boolean isActive;
}